+++ src/serial.rs
@@ -1743,15 +1743,15 @@ macro_rules! halUsart {
         $(
         impl<PINS> Serial<$USARTX, PINS> {
             fn config_stop(self, config: config::Config) -> Self {
-                use crate::stm32::usart1::cr2::STOPW;
+                use crate::stm32::usart1::cr2::STOP_A;
                 use self::config::*;
 
                 self.usart.cr2.write(|w| {
                     w.stop().variant(match config.stopbits {
-                        StopBits::STOP0P5 => STOPW::STOP0P5,
-                        StopBits::STOP1 => STOPW::STOP1,
-                        StopBits::STOP1P5 => STOPW::STOP1P5,
-                        StopBits::STOP2 => STOPW::STOP2,
+                        StopBits::STOP0P5 => STOP_A::STOP0P5,
+                        StopBits::STOP1 => STOP_A::STOP1,
+                        StopBits::STOP1P5 => STOP_A::STOP1P5,
+                        StopBits::STOP2 => STOP_A::STOP2,
                     })
                 });
                 self
@@ -1785,15 +1785,15 @@ macro_rules! halUart {
         $(
         impl<PINS> Serial<$USARTX, PINS> {
             fn config_stop(self, config: config::Config) -> Self {
-                use crate::stm32::uart4::cr2::STOPW;
+                use crate::stm32::uart4::cr2::STOP_A;
                 use self::config::*;
 
                 self.usart.cr2.write(|w| {
                     w.stop().variant(match config.stopbits {
-                        StopBits::STOP0P5 => STOPW::STOP1,
-                        StopBits::STOP1 => STOPW::STOP1,
-                        StopBits::STOP1P5 => STOPW::STOP2,
-                        StopBits::STOP2 => STOPW::STOP2,
+                        StopBits::STOP0P5 => STOP_A::STOP1,
+                        StopBits::STOP1 => STOP_A::STOP1,
+                        StopBits::STOP1P5 => STOP_A::STOP2,
+                        StopBits::STOP2 => STOP_A::STOP2,
                     })
                 });
                 self