+++ examples/stopwatch-with-ssd1306-and-interrupts.rs
@@ -14,6 +14,7 @@ use crate::hal::{
     prelude::*,
     rcc::{Clocks, Rcc},
     stm32,
+    syscfg,
     timer::{Event, Timer},
 };
 use arrayvec::ArrayString;
@@ -44,8 +45,6 @@ enum StopwatchState {
 #[entry]
 fn main() -> ! {
     if let (Some(mut dp), Some(cp)) = (stm32::Peripherals::take(), cortex_m::Peripherals::take()) {
-        dp.RCC.apb2enr.write(|w| w.syscfgen().enabled());
-
         let rcc = dp.RCC.constrain();
         let clocks = setup_clocks(rcc);
         let gpiob = dp.GPIOB.split();
@@ -59,10 +58,13 @@ fn main() -> ! {
             clocks,
         );
 
+        // Enable the system configuration controller
+        let mut syscfg = syscfg::Syscfg::new(dp.SYSCFG);
+
         // Create a button input with an interrupt
         let gpioc = dp.GPIOC.split();
         let mut board_btn = gpioc.pc13.into_pull_up_input();
-        board_btn.make_interrupt_source(&mut dp.SYSCFG);
+        board_btn.make_interrupt_source(&mut syscfg);
         board_btn.enable_interrupt(&mut dp.EXTI);
         board_btn.trigger_on_edge(&mut dp.EXTI, Edge::FALLING);