+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Added examples in the examples folder.
+- [breaking-change] Added a wrapper around SYSCFG.
 
 ## [v0.6.0] - 2019-10-19